refactor(pytest-bdd): merge utils
