+++ allure-pytest-bdd/src/plugin.py
@@ -5,13 +5,14 @@ import allure_commons
 from allure_commons.logger import AllureFileLogger
 from allure_commons.lifecycle import AllureLifecycle
 
-from .api import AllurePytestBddApi
+from .hooks import AllurePytestBddApiHooks
 from .pytest_bdd_listener import PytestBDDListener
-from .api import ALLURE_TITLE_MARK
-from .api import ALLURE_DESCRIPTION_MARK
-from .api import ALLURE_DESCRIPTION_HTML_MARK
-from .api import ALLURE_LABEL_MARK
-from .api import ALLURE_LINK_MARK
+
+from .utils import ALLURE_TITLE_MARK
+from .utils import ALLURE_DESCRIPTION_MARK
+from .utils import ALLURE_DESCRIPTION_HTML_MARK
+from .utils import ALLURE_LABEL_MARK
+from .utils import ALLURE_LINK_MARK
 
 
 def pytest_addoption(parser):
@@ -80,7 +81,7 @@ def pytest_configure(config):
         allure_commons.plugin_manager.register(pytest_bdd_listener)
         config.add_cleanup(cleanup_factory(pytest_bdd_listener))
 
-        allure_api_impl = AllurePytestBddApi(config, lifecycle)
+        allure_api_impl = AllurePytestBddApiHooks(config, lifecycle)
         allure_commons.plugin_manager.register(allure_api_impl)
         config.add_cleanup(cleanup_factory(allure_api_impl))
