+++ tests/allure_pytest_bdd/acceptance/labels/features_test.py
@@ -75,7 +75,6 @@ def test_dynamic_feature(allure_pytest_bdd_runner: AllurePytestRunner):
     allure_results = allure_pytest_bdd_runner.run_pytest(
         ("sample.feature", feature_content),
         steps_content,
-        cli_args=["--capture=no"]
     )
 
     assert_that(
